+++ installation_instruction/__main__.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 from sys import exit
-from os.path import isfile
+from os.path import isfile, isdir
 from subprocess import run
 import platform
 
@@ -22,7 +22,8 @@ import click
 from .__init__ import __version__, __description__, __repository__, __author__, __author_email__, __license__
 from .get_flags_and_options_from_schema import _get_flags_and_options
 from .installation_instruction import InstallationInstruction
-from .helpers import _make_pretty_print_line_breaks
+from .helpers import _make_pretty_print_line_breaks, _is_remote_git_repository, _clone_git_repo, _config_file_is_in_folder
+
 
 VERSION_STRING = f"""Version: installation-instruction {__version__}
 Copyright: (C) 2024 {__author_email__}, {__author__}
@@ -46,30 +47,52 @@ def _get_system(option_types):
 
     return None
 
+def _red_echo(text: str):
+    click.echo(click.style(text, fg="red"))
+
+
 class ConfigReadCommand(click.MultiCommand):
     """
-    Custom click command class to read config file and show installation instructions with parameters.
+    Custom click command class to read config file, folder or git repository and show installation instructions with parameters.
     """
 
     def __init__(self, *args, **kwargs):
         super().__init__(
             *args,
             **kwargs,
-            subcommand_metavar="CONFIG_FILE [OPTIONS]...",
+            subcommand_metavar="CONFIG_FILE/FOLDER/GIT_REPO_URL [OPTIONS]...",
             options_metavar="",
         )
 
 
     def get_command(self, ctx, config_file: str) -> click.Command|None:
-        if not isfile(config_file):
-            click.echo("Config file not found.")
-            return None
 
+        temp_dir = None
+        if _is_remote_git_repository(config_file):
+            try:
+                temp_dir = _clone_git_repo(config_file)
+            except Exception as e:
+                _red_echo("Error (cloning git repository):\n\n" + str(e))
+                exit(1)
+            config_file = temp_dir.name
+        if isdir(config_file):
+            if path := _config_file_is_in_folder(config_file):
+                config_file = path
+            else:
+                if temp_dir is not None:
+                    _red_echo("Config file not found in repository.")
+                else:
+                    _red_echo(f"Config file not found in folder {config_file}")
+                exit(1)
+        if not isfile(config_file):
+            _red_echo(f"{config_file} is not a file.")
+            exit(1)
+        
         try:
             instruction = InstallationInstruction.from_file(config_file)
-            options = _get_flags_and_options(instruction.schema)
+            options = _get_flags_and_options(instruction.schema, getattr(instruction, "misc", None))
         except Exception as e:
-            click.echo(click.style("Error (parsing options from schema): " + str(e), fg="red"))
+            _red_echo("Error (parsing options from schema): " + str(e))
             exit(1)
 
         for option in options:
@@ -81,7 +104,7 @@ class ConfigReadCommand(click.MultiCommand):
         def callback(**kwargs):
             inst = instruction.validate_and_render(kwargs)
             if inst[1]:
-                click.echo(click.style("Error: " + inst[0], fg="red"))
+                _red_echo("Error: " + inst[0])
                 exit(1)
             if ctx.obj["MODE"] == "show":
                 if ctx.obj["RAW"]:
@@ -91,7 +114,7 @@ class ConfigReadCommand(click.MultiCommand):
             elif ctx.obj["MODE"] == "install":
                 result = run(inst[0], shell=True, text=True, capture_output=True)
                 if result.returncode != 0:
-                    click.echo(click.style("Installation failed with:\n" + str(result.stdout) + "\n" + str(result.stderr), fg="red"))
+                    _red_echo("Installation failed with:\n" + str(result.stdout) + "\n" + str(result.stderr))
                     exit(1)
                 else:
                     if ctx.obj["INSTALL_VERBOSE"]:
@@ -106,7 +129,6 @@ class ConfigReadCommand(click.MultiCommand):
             params=options,
             callback=callback,
         )
-
 
 @click.command(cls=ConfigReadCommand, help="Shows installation instructions for your specified config file and parameters.")
 @click.option("--raw", is_flag=True, help="Show installation instructions without pretty print.", default=False)
@@ -133,3 +155,4 @@ main.add_command(install)
 
 if __name__ == "__main__":
     main()
+